refactor: 자동차의 이동전략을 인터페이스를 이용해 전략 패턴으로 구현하라
- 현재는 랜덤값을 이용해 이동하는 전략이지만 이동전략은 변경될 여지가 있기때문에 변경에 유연한 전략 패턴을 이용한다
