+++ src/main/java/racingcar/domain/strategy/MoveStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.domain.strategy;
+
+public interface MoveStrategy {
+    boolean isMovable();
+}
